+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".auth.AuthActivity">
+        <activity android:name=".auth.views.AuthActivity">
         </activity>
     </application>
