+++ app/src/main/java/com/example/draftt/auth/views/AuthActivity.kt
@@ -1,4 +1,4 @@
-package com.example.draftt.auth
+package com.example.draftt.auth.views
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
